Half animal can be packed
The half animal can now be packed correctly. To do this, I had to work from the bottom up and adjust how the animals were registered, cut up, and placed into trays. There is some persistence in the system, but it is still missing the checks for if parts are used. The one to many mapping had to be changed and the join tables had to be specified as opposed to assumed.