Extraction and Used checks
I extracted the methods for building the responses all the way from the animal message to the product list. In addition, I worked on checking if things were used and sending detailed exceptions back for some of them. I can see that the logic doesn't work perfectly on the half animal, but it looks good elsewhere.